+++ data/macros/drink-dirty-water.js
@@ -1,0 +1,14 @@
+/**
+ ***************************************************************
+ * This macro can be used to have a character drink dirty water
+ * from the environment rather than by using a Dirty Water
+ * consumable item from their inventory.
+ *
+ * When a GM runs this macro, it will use the currently
+ * selected character token.
+ *
+ * When a player runs this macro, it will use their assigned
+ * primary character.
+ **************************************************************/
+
+fallout.macros.drinkDirtyWater();